Move profile page into its own module
